added new api in admin getdataperroundbygroupnoandtipmode

+++ CleanArchitecture.Application/Models/Admin/SummaryAddDataResponse.cs
@@ -32,7 +32,7 @@ namespace ComplexCalculator.Application.Models.Admin
     public class ScoreBoardGrid
     {
         public float MainTube { get; set; } = 0;
-        public int Luozi { get; set; } = 0;
+        public float Luozi { get; set; } = 0;
         public float SmallTable { get; set; } = 0;
         public float Statistics { get; set; } = 0;
         public int InputByUser { get; set; }
